[FIX] - Fix cast in 'i', L.25

+++ 0x15-file_io/1-create_file.c
@@ -22,7 +22,7 @@ int create_file(const char *filename, char *text_content)
 		i++;
 	}
 
-	val = write(file, text_content, i);
+	val = write(file, text_content, (size_t)i);
 	if ((int)val == -1)
 		return (-1);
